+++ lib/services/auth/auth_provider.dart
@@ -19,7 +19,13 @@ abstract class AuthProvider{
       required String password,
     }
   );
+  Future<void> sentResetLink(
+    {
+      required String email,
+    }
+  );
   Future<void> logOut();
   Future<void> sendEmailVerification();
+  
 
 }
