+++ lib/src/constants/text_strings.dart
@@ -1,0 +1,18 @@
+//Login text
+const String tEmail = 'Email';
+const String tPhone = 'Phone Number';
+
+
+//Forgot password text
+const String tForgetPasswordTitle = 'Forgot Password?';
+const String tForgetPswrdTitle = 'Select option';
+const String tForgetPswrdSubTitle = 'Select one of the below method to verify your account.';
+const String tResetviaEmail = 'Reset password via email address';
+const String tResetViaPhone = 'Reset password via phone number';
+const String tResetUsingEmail = 'Enter your email and we will send a password reset link to you.';
+const String tResetBtn = 'Reset password';
+const String tEmailNotFound = 'User not found';
+const String tLinkSent = 'Password reset link sent! Check your email';
+
+//Padding
+const double tPaddingSize = 20.0;
