+++ lib/src/features/auth/screens/forgetPswrd/forget_pswrd_btn.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+
+class ForgetPasswordWidget extends StatelessWidget {
+  const ForgetPasswordWidget({
+    required this.btnIcon,
+    required this.title,
+    required this.subTitle,
+    required this.onTap,
+    super.key,
+  });
+
+  final IconData btnIcon;
+  final String title, subTitle;
+  final VoidCallback onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        padding: const EdgeInsets.all(20),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(10),
+          color: const Color.fromARGB(255, 231, 231, 231),
+        ),
+        child: Row(
+          children: [
+            Icon(btnIcon, size: 50),
+            const SizedBox(width: 20),
+            Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  title,
+                  style: TextStyle(fontSize: 19, fontWeight: FontWeight.bold),),
+                Text(
+                  subTitle,
+                  style: TextStyle(fontSize: 15),),
+              ],
+            )
+          ]),
+      ),
+    );
+  }
+}
