+++ lib/src/features/users/customer/cust_profile.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/constants/routes.dart';
+import 'package:flutter_application_1/src/routing/routes_const.dart';
 import 'package:flutter_application_1/services/auth/auth_service.dart';
 
 class CustomerProfilePage extends StatefulWidget {
