+++ lib/src/routing/routes_const.dart
@@ -10,3 +10,4 @@ const deliveryManOwnerRoute = '/delivery_homepage/';
 const custProfileRoute = '/customer profile/';
 const editProfileRoute = '/edit profile/';
 const testReadRoute = '/test_read/';
+const resetPswrdEmailRoute = '/forget_pswrd_email/';
