+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   path_provider: ^2.1.1
   path: ^1.8.3
   cloud_firestore: ^4.12.2
+  get: ^4.6.6
 
 dev_dependencies:
   flutter_test:
